// functions/api/documents/model/DocumentInfo.ts
export interface DocumentInfo {
  documentId: string
  pageTitle: string
  pageData: string
  editLock: boolean
  archive: boolean
  categoryId: string
  categoryName: string
  pageRegisterUserId: string
  pageRegisterDate: string
  pageUpdateUserId: string
  pageUpdateDate: string
  pageVersion: number
}